requirements.txt
lanchain_openai
langchain_core
python-dotenv
streamlit 
langchain_community